requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
anyio==4.0.0
apturl==0.5.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.12.2
blinker==1.4
Brlapi==0.7.0
cachetools==5.3.2
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.3.2
chromadb==0.3.29
Click==7.0
clickhouse-connect==0.6.18
colorama==0.4.3
coloredlogs==15.0.1
command-not-found==0.3
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
dataclasses-json==0.5.14
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
dnspython==2.4.2
docarray==0.39.0
duckdb==0.9.1
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.1.3
fastapi==0.85.1
fasteners==0.14.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.2
gitdb==4.0.11
GitPython==3.1.40
graphlib-backport==1.0.3
greenlet==3.0.1
h11==0.14.0
hnswlib==0.7.0
httplib2==0.14.0
httptools==0.6.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==2.8
importlib-metadata==6.8.0
importlib-resources==6.1.1
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
keyring==18.0.1
kiwisolver==1.4.5
langchain==0.0.212
langchainplus-sdk==0.0.20
language-selector==0.1
lark==1.1.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
loguru==0.7.2
louis==3.12.0
lz4==4.3.2
macaroonbakery==1.3.1
Mako==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.1
mdurl==0.1.2
monotonic==1.5
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.10.4
numexpr==2.8.6
numpy==1.24.3
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.16.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pandas==2.0.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==9.5.0
pinecone-client==2.2.4
pkgutil_resolve_name==1.3.10
posthog==3.0.2
protobuf==4.24.4
pulsar-client==3.3.0
pyarrow==14.0.0
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.16.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.1.1
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.0
pytz==2023.3.post1
pyxdg==0.26
PyYAML==6.0
referencing==0.30.2
regex==2023.10.3
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
rich==13.6.0
rpds-py==0.12.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.20.4
streamlit==1.27.2
sympy==1.12
systemd-python==234
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
types-requests==2.31.0.10
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.7
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.22.0
wadllib==1.3.3
watchdog==3.0.0
watchfiles==0.21.0
websockets==12.0
xkit==0.0.0
yarl==1.9.2
zipp==3.17.0
zstandard==0.22.0
